scraper/requirements.txt
scrapy==2.11.2
scrapy-user-agents==0.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1